[Core] Move old fixtures to CoreBundle
