+++ src/Sylius/Bundle/CoreBundle/Resources/fixtures/mugs.jpg
