+++ src/Sylius/Bundle/CoreBundle/Resources/fixtures/stickers.jpg
