+++ src/Sylius/Bundle/CoreBundle/Resources/fixtures/t-shirts.jpg
